requirements.txt
openai>=1.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
